requirements.txt
aiogram==2.20
python-dotenv==0.20.0
requests==2.27.1
